requirements.txt
python-telegram-bot==20.7
feedparser
yt-dlp
requests
python-dotenv